+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### TBD
+- Add `moveRows` to `Query` module, along with related types: `MoveRowsOptions` and `MoveRowsResponse`
+
 ### 1.27.1 - 2023-11-9
 - Fix `Utils.encodeHtml` by adding necessary semicolon